% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_dynamics.R
\name{bi_exponential_decay}
\alias{bi_exponential_decay}
\title{Simulate bi-exponential decay}
\usage{
bi_exponential_decay(t, A0, par_sigma, par_gamma, par_rho)
}
\arguments{
\item{t}{Time}

\item{A0}{Peak}

\item{par_sigma}{Rate of short-term decay}

\item{par_gamma}{Rate of long-term decay}

\item{par_rho}{Proportion of short-term decay}
}
\value{
A numeric vector
}
\description{
Simulate bi-exponential decay
}
\examples{
bi_exponential_decay(365, exp(10), 0.98, 1/720, 1/30)
}
