% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/before_enrolment.R
\name{estimate_lambda_e}
\alias{estimate_lambda_e}
\title{Estimate the cumulative force of infection at enrolment}
\usage{
estimate_lambda_e(n_e, yearly_lambda, start_index, init_weight, final_weight)
}
\arguments{
\item{n_e}{A non-negative integer.}

\item{yearly_lambda}{A numeric vector}

\item{start_index}{A positive integer indicating the year in which the
individual became susceptible after birth.}

\item{init_weight}{A numeric value corresponding to the fraction of the first
year in which the individual was susceptible.}

\item{final_weight}{A numeric value representing the fraction of the calendar
year from the 1st Jan to the date of enrolment}
}
\value{
A real number
}
\description{
Estimate the cumulative force of infection at enrolment
}
\examples{
estimate_lambda_e(3, rep(0.1, 26), 1, 0.25, 0.75)
}
