% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOI.R
\name{random_walk_lambda}
\alias{random_walk_lambda}
\title{Simulate a random walk using Geometric Brownian Motion with no drift}
\usage{
random_walk_lambda(n_t, dt, sigma, x0)
}
\arguments{
\item{n_t}{An integer indicating the number of years}

\item{dt}{A real number indicating the time step for the Euler-Maruyama algorithm}

\item{sigma}{A real number corresponding to the volatility parameter}

\item{x0}{A real number for the initial value}
}
\value{
a numeric vector
}
\description{
Simulate a random walk using Geometric Brownian Motion with no drift
}
\examples{
random_walk_lambda(26, 0.01, 0.15, 0.1)
}
