% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infections_since_birth.R
\name{simulate_infections_since_birth}
\alias{simulate_infections_since_birth}
\title{Simulate infections since birth}
\usage{
simulate_infections_since_birth(
  individuals_df,
  lambda,
  rho,
  stop_index,
  vac_buckets = 0,
  rho_v = 0,
  is_vac_prob = 0
)
}
\arguments{
\item{individuals_df}{A data frame}

\item{lambda}{A numeric vector}

\item{rho}{A numeric value}

\item{stop_index}{An integer}

\item{vac_buckets}{An integer}

\item{rho_v}{A numeric value}

\item{is_vac_prob}{A boolean.}
}
\value{
A data frame
}
\description{
Simulate infections since birth
}
\examples{
individuals_df <- data.frame(subject_id       = 1,
                             birth_year_index = 9,
                             is_vaccinated    = 1,
                             vac_year         = 15)


simulate_infections_since_birth(individuals_df = individuals_df,
                                lambda         = rep(0.14, 15),
                                rho            = 0.05,
                                stop_index     = 15,
                                vac_buckets    = 1,
                                rho_v          = 0.12,
                                is_vac_prob    = 1)
}
