% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_dynamics.R
\name{add_titre_to_infection_history}
\alias{add_titre_to_infection_history}
\title{Add titres to infection history}
\usage{
add_titre_to_infection_history(
  inf_history_df,
  long_decay_coefficient,
  peak_scaling_factor
)
}
\arguments{
\item{inf_history_df}{A data frame}

\item{long_decay_coefficient}{A scalar in the interval (0, 1)
that modulates the long-term decay rate, representing the proportion of the
long-term rate that is effectively applied to the system.}
}
\value{
A data frame
}
\description{
Add titres to infection history
}
\examples{
inf_history_df <-
  data.frame(subject_id = 1,
             year_index = 11:23,
             sim_y      = 0)

inf_history_df$sim_y[c(6, 8, 11, 12)] <- 1

add_titre_to_infection_history(inf_history_df)
}
