% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VE.R
\name{estimate_cumulative_VE}
\alias{estimate_cumulative_VE}
\title{Title}
\usage{
estimate_cumulative_VE(follow_up_df)
}
\arguments{
\item{follow_up_df}{A data frame}
}
\value{
A data frame
}
\description{
Title
}
\examples{
  follow_up_df <- data.frame(subject_id    = rep(1:4, 5),
                             year_index    = rep(1:5, each = 4),
                             is_vaccinated = rep(c(1, 1, 0, 0), 5),
                             infection     = c(0, 0, 1, 0,
                                               0, 1, 0, 1,
                                               1, 0, 0, 0,
                                               0, 0, 0, 0,
                                               0, 0, 0, 0))
 estimate_cumulative_VE(follow_up_df)
}
