% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infections_since_birth.R
\name{simulate_infections_since_birth}
\alias{simulate_infections_since_birth}
\title{Simulate infections since birth}
\usage{
simulate_infections_since_birth(individuals_df, lambda, rho, stop_index)
}
\arguments{
\item{individuals_df}{A data frame}

\item{lambda}{A numeric vector}

\item{rho}{A numeric value}

\item{stop_index}{An integer}
}
\value{
A data frame
}
\description{
Simulate infections since birth
}
\examples{
individuals_df <- data.frame(subject_id       = 1,
                             birth_year_index = 9,
                             is_vaccinated    = 1)

rho        <- 0.05
stop_index <- 15
lambda     <- rep(0.14, 15)

simulate_infections_since_birth(individuals_df, lambda, rho, stop_index)
}
